通知機能.

+++ app/assets/javascripts/edit_place.js
@@ -1,21 +1,22 @@
 $(document).on("turbolinks:load", function(){
   $('.delete-btn').on('click', function(e) {
     e.preventDefault()
-    alert('画像を削除します。\nよろしいですか？')
-    var Click = $(this);
-    var image_id = $(this).data('imgid');
-    $.ajax({
-      type: "DELETE",
-      url: '/api/destroy/image_destroy',
-      data: {img_id: image_id},
-      dataType: 'json'
-    })
-    .done(function() {
-      Click.parent('.img_view').next('input[type="hidden"]').val('');
-      Click.parent().remove();
-    })
-    .fail(function() {
-      alert('削除に失敗しました');
-    })
+    if(window.confirm('画像を削除します。\nよろしいですか？')){
+      var Click = $(this);
+      var image_id = $(this).data('imgid');
+      $.ajax({
+        type: "DELETE",
+        url: '/api/destroy/image_destroy',
+        data: {img_id: image_id},
+        dataType: 'json'
+      })
+      .done(function() {
+        Click.parent('.img_view').next('input[type="hidden"]').remove('');
+        Click.parent().remove();
+      })
+      .fail(function() {
+        alert('削除に失敗しました');
+      })
+    }
   });
 });
